requirements.txt
torch
torchvision
codecarbon
numpy
pandas
seaborn
matplotlib
einops
tqdm
opencv-python
scikit-image
scipy
scikit-learn
PyYAML